refactor(new recipe page): split code
