+++ src/app/components/NewRecipe/style.module.scss
