+++ src/app/recipes/new/components/SearchProductModal/index.tsx
@@ -23,7 +23,7 @@ import IngredientInformationHeader from '@/app/recipes/[recipeKey]/components/In
 
 import NewProduct from './NewProduct';
 import ExistingProduct from './ExistingProduct';
-import { NewRecipeIngredientState } from '../../NewRecipe';
+import { NewRecipeIngredientState } from '../../../../components/NewRecipe';
 import { INGREDIENT_TEXT_LIMIT, OnSelectProductProps } from '../Ingredients';
 import { getProducts } from '@/services/products';
